数据库逻辑细化完成，数据库文件放入resources文件夹，剩下说明文档

+++ src/main/java/com/KTP/controller/studentQuitCourseController.java
@@ -1,6 +1,8 @@
 package com.KTP.controller;
 
+import com.KTP.service.courseService;
 import com.KTP.service.courseStudentService;
+import com.KTP.service.signService;
 import com.KTP.service.studentService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -25,6 +27,18 @@ public class studentQuitCourseController {
         this.courseStudentService = courseStudentService;
     }
 
+    private signService signService;
+    @Autowired
+    public void setSignService(signService signService){
+        this.signService = signService;
+    }
+
+    private courseService courseService;
+    @Autowired
+    public void setCourseService(courseService courseService){
+        this.courseService = courseService;
+    }
+
     @RequestMapping(path = "/studentQuitCourse")
     public String toStudentQuitCourse(HttpServletRequest request, Model model){
         String sno = (String)request.getSession().getAttribute("sno");
@@ -38,15 +52,22 @@ public class studentQuitCourseController {
         String sno = (String)request.getSession().getAttribute("sno");
         String sname = studentService.getStudentAdminSname(sno);
         model.addAttribute("sname", sname);
-        String cno = courseStudentService.getCourseStudentCno(sno, kh);
+        String cno = courseService.getCourseCno(kh);
+        String snocheck = courseStudentService.getCourseStudentCno(sno, kh);
         if (cno == null){
             request.setAttribute("message", "您输入的课程号不存在！");
         }
         else {
-            try {
-                courseStudentService.deleteCourseStudent(sno, kh);
-                request.setAttribute("message", "退课成功！");
-            }catch (Exception e){
+            if (snocheck != null){
+                try {
+                    courseStudentService.deleteCourseStudent(sno, kh);
+                    signService.deleteSignStudent(kh, sno);
+                    request.setAttribute("message", "退课成功！");
+                }catch (Exception e){
+                    request.setAttribute("message", "退课失败！");
+                }
+            }
+            else {
                 request.setAttribute("message", "您未加入该课程！");
             }
         }
